refactored, removed references to optimize, clean up and updated tests
